requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3
imbalanced-learn==0.11.0

# Explainability
shap==0.42.1

# API & Web
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0

# Data Generation
faker==19.6.2

# Visualization & Analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2

# Testing
pytest==7.4.2
httpx==0.24.1

# Monitoring
evidently==0.4.8

# Development
jupyter==1.0.0
ipykernel==6.25.2
black==23.9.1
